Refactor: Split LLMClientBase into LLMClientBase and ChatClientBase.
