+++ src/floki/__init__.py
@@ -3,7 +3,7 @@ from floki.agent import (
     AgenticWorkflowService, RoundRobinWorkflowService, RandomWorkflowService,
     LLMWorkflowService, ReActAgent, ToolCallAgent, OpenAPIReActAgent
 )
-from floki.llm import LLMClientBase
+from floki.llm import LLMClientBase, ChatClientBase
 from floki.llm.openai import OpenAIChatClient
 from floki.llm.huggingface import HFHubChatClient
 from floki.tool import AgentTool, tool
